requirements.txt
tensorflow==2.20.0
transformers>=4.45.1
langchain>=0.2.16
langchain-openai>=0.1.20
langchain-core>=0.2.38
langchain-text-splitters>=0.2.0
langsmith>=0.1.17
pydantic>=2.0
pyyaml>=6.0
sqlalchemy>=2.0
tenacity>=8.2.0
aiohttp>=3.9.0
filelock>=3.12.0
huggingface-hub>=0.20.0
regex>=2023.6.0
safetensors>=0.4.0
tokenizers>=0.15.0
tqdm>=4.66.0
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
openai==1.51.0
torch>=2.4.0